(feat) refactor components with a base class
