+++ hummingbot/data_feed/candles_feed/binance_spot_candles/binance_spot_candles.py
@@ -1,63 +1,28 @@
 import asyncio
 import logging
-from collections import deque
 from typing import Any, Dict, Optional
 
 import numpy as np
-import pandas as pd
 
-from hummingbot.core.api_throttler.async_throttler import AsyncThrottler
-from hummingbot.core.network_base import NetworkBase
 from hummingbot.core.network_iterator import NetworkStatus
-from hummingbot.core.utils.async_utils import safe_ensure_future
 from hummingbot.core.web_assistant.connections.data_types import WSJSONRequest
-from hummingbot.core.web_assistant.web_assistants_factory import WebAssistantsFactory
 from hummingbot.core.web_assistant.ws_assistant import WSAssistant
 from hummingbot.data_feed.candles_feed.binance_spot_candles import constants as CONSTANTS
+from hummingbot.data_feed.candles_feed.candles_base import CandlesBase
 from hummingbot.logger import HummingbotLogger
 
 
-class BinanceSpotCandles(NetworkBase):
-    _bcf_logger: Optional[HummingbotLogger] = None
-    _binance_candles_shared_instance: "BinanceSpotCandles" = None
-    columns = ["timestamp", "open", "low", "high", "close", "volume", "quote_asset_volume",
-               "n_trades", "taker_buy_base_volume", "taker_buy_quote_volume"]
+class BinanceSpotCandles(CandlesBase):
+    _logger: Optional[HummingbotLogger] = None
 
     @classmethod
     def logger(cls) -> HummingbotLogger:
-        if cls._bcf_logger is None:
-            cls._bcf_logger = logging.getLogger(__name__)
-        return cls._bcf_logger
+        if cls._logger is None:
+            cls._logger = logging.getLogger(__name__)
+        return cls._logger
 
     def __init__(self, trading_pair: str, interval: str = "1m", max_records: int = 150):
-        super().__init__()
-        async_throttler = AsyncThrottler(rate_limits=self.rate_limits)
-        self._api_factory = WebAssistantsFactory(throttler=async_throttler)
-
-        self._trading_pair = trading_pair
-        self._ex_trading_pair = trading_pair.replace("-", "")
-        if interval in CONSTANTS.INTERVALS.keys():
-            self._interval = interval
-        else:
-            self.logger().exception(f"Interval {interval} is not supported. Available Intervals: {CONSTANTS.INTERVALS.keys()}")
-            raise
-
-        self._candles = deque(maxlen=max_records)
-        self._listen_candles_task: Optional[asyncio.Task] = None
-        self.start()
-
-    async def start_network(self):
-        await self.stop_network()
-        self._listen_candles_task = safe_ensure_future(self.listen_for_subscriptions())
-
-    async def stop_network(self):
-        if self._listen_candles_task is not None:
-            self._listen_candles_task.cancel()
-            self._listen_candles_task = None
-
-    @property
-    def is_ready(self):
-        return len(self._candles) == self._candles.maxlen
+        super().__init__(trading_pair, interval, max_records)
 
     @property
     def name(self):
@@ -66,6 +31,10 @@ class BinanceSpotCandles(NetworkBase):
     @property
     def rest_url(self):
         return CONSTANTS.REST_URL
+
+    @property
+    def wss_url(self):
+        return CONSTANTS.WSS_URL
 
     @property
     def health_check_url(self):
@@ -89,9 +58,8 @@ class BinanceSpotCandles(NetworkBase):
                                              throttler_limit_id=CONSTANTS.HEALTH_CHECK_ENDPOINT)
         return NetworkStatus.CONNECTED
 
-    @property
-    def candles(self) -> pd.DataFrame:
-        return pd.DataFrame(self._candles, columns=self.columns, dtype=float)
+    def get_exchange_trading_pair(self, trading_pair):
+        return trading_pair.replace("-", "")
 
     async def fetch_candles(self,
                             start_time: Optional[int] = None,
@@ -127,35 +95,6 @@ class BinanceSpotCandles(NetworkBase):
                     "Unexpected error occurred when getting historical klines. Retrying in 1 seconds...",
                 )
                 await self._sleep(1.0)
-
-    async def listen_for_subscriptions(self):
-        """
-        Connects to the trade events and order diffs websocket endpoints and listens to the messages sent by the
-        exchange. Each message is stored in its own queue.
-        """
-        ws: Optional[WSAssistant] = None
-        while True:
-            try:
-                ws: WSAssistant = await self._connected_websocket_assistant()
-                await self._subscribe_channels(ws)
-                await self._process_websocket_messages(websocket_assistant=ws)
-            except asyncio.CancelledError:
-                raise
-            except ConnectionError as connection_exception:
-                self.logger().warning(f"The websocket connection was closed ({connection_exception})")
-            except Exception:
-                self.logger().exception(
-                    "Unexpected error occurred when listening to public klines. Retrying in 1 seconds...",
-                )
-                await self._sleep(1.0)
-            finally:
-                await self._on_order_stream_interruption(websocket_assistant=ws)
-
-    async def _connected_websocket_assistant(self) -> WSAssistant:
-        ws: WSAssistant = await self._api_factory.get_ws_assistant()
-        await ws.connect(ws_url=CONSTANTS.WSS_URL,
-                         ping_timeout=30)
-        return ws
 
     async def _subscribe_channels(self, ws: WSAssistant):
         """
@@ -207,13 +146,3 @@ class BinanceSpotCandles(NetworkBase):
                     self._candles.append(np.array([timestamp, open, low, high, close, volume,
                                                    quote_asset_volume, n_trades, taker_buy_base_volume,
                                                    taker_buy_quote_volume]))
-
-    async def _sleep(self, delay):
-        """
-        Function added only to facilitate patching the sleep in unit tests without affecting the asyncio module
-        """
-        await asyncio.sleep(delay)
-
-    async def _on_order_stream_interruption(self, websocket_assistant: Optional[WSAssistant] = None):
-        websocket_assistant and await websocket_assistant.disconnect()
-        self._candles.clear()
