+++ test/hummingbot/data_feed/candles_feed/binance_perpetuals_candles/test_binance_perpetuals_candles.py
@@ -288,7 +288,7 @@ class TestBinancePerpetualsCandles(unittest.TestCase):
 
         self.listening_task = self.ev_loop.create_task(self.data_feed.listen_for_subscriptions())
 
-        self.mocking_assistant.run_until_all_aiohttp_messages_delivered(ws_connect_mock.return_value)
+        self.mocking_assistant.run_until_all_aiohttp_messages_delivered(ws_connect_mock.return_value, timeout=2)
 
         self.assertEqual(self.data_feed.candles.shape[0], 1)
         self.assertEqual(self.data_feed.candles.shape[1], 10)
